+++ NEWS.md
@@ -14,3 +14,4 @@
 * Changed the default number of CPU cores employed by the estimation function fitGMVAR to be at most two due to CRAN policy.
 * Added argument "seeds" to fitGMVAR allowing one to set the random number generator seed for each call to the genetic algorithm.
 * New exported function: alt_gmvar
+* Updated the vignette to include the definition of the GMVAR model