+++ NEWS.md
@@ -15,3 +15,8 @@
 * Added argument "seeds" to fitGMVAR allowing one to set the random number generator seed for each call to the genetic algorithm.
 * New exported function: alt_gmvar
 * Updated the vignette to include the definition of the GMVAR model
+
+# gmvarkit 1.1.2
+
+* Improvements on the comments and documentation.
+* Minor update on the genetic algorithm.